# frontend/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Asclepius",
    "description": "Self-hosted medical records manager",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/settings/view-prefs/{view_key}": {
      "get": {
        "tags": ["settings"],
        "summary": "Get View Prefs",
        "description": "Return the user's stored prefs for ``view_key`` or ``null`` keys\nwhen no row exists. Frontend treats null as \"use defaults\".",
        "operationId": "get_view_prefs_api_settings_view_prefs__view_key__get",
        "parameters": [
          {
            "name": "view_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "View Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["settings"],
        "summary": "Set View Prefs",
        "operationId": "set_view_prefs_api_settings_view_prefs__view_key__put",
        "parameters": [
          {
            "name": "view_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "View Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewPrefsBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/vault/tree": {
      "get": {
        "tags": ["vault"],
        "summary": "Get Vault Tree",
        "description": "Get the vault directory structure as a browsable tree.",
        "operationId": "get_vault_tree_api_vault_tree_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Subtree path relative to vault root",
              "title": "Path"
            },
            "description": "Subtree path relative to vault root"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/vault/file": {
      "delete": {
        "tags": ["vault"],
        "summary": "Delete Vault File",
        "description": "Delete a file from the vault and its matching document record.",
        "operationId": "delete_vault_file_api_vault_file_delete",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "File path relative to vault root",
              "title": "Path"
            },
            "description": "File path relative to vault root"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/vault/move": {
      "post": {
        "tags": ["vault"],
        "summary": "Move Vault File",
        "description": "Move a file (or imaging study folder) within the vault and keep the\nmatching document record's ``file_path`` in sync.\n\nUse case: a file was misclassified by date / event and the user wants\nto drag it into the right folder without losing the document reference\n(so document-detail links, OCR cache, lab results etc. all keep\nworking). The DB row is updated atomically with the disk move; if the\ndestination already exists or the access check fails the move is\nrejected.",
        "operationId": "move_vault_file_api_vault_move_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoveRequest": {
        "properties": {
          "from_path": {
            "type": "string",
            "title": "From Path"
          },
          "to_path": {
            "type": "string",
            "title": "To Path"
          }
        },
        "type": "object",
        "required": ["from_path", "to_path"],
        "title": "MoveRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
      "ViewPrefsBody": {
        "properties": {
          "visible": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Visible"
          },
          "order": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Order"
          }
        },
        "type": "object",
        "title": "ViewPrefsBody"
      }
    }
  }
}